Added Put to BooksFOrmCOntroller

+++ src/main/java/com/codecool/BookShop/controller/BooksFormController.java
@@ -31,20 +31,31 @@ public class BooksFormController {
             return bookform.get();
         }
 
+//        @DeleteMapping("/bookForm/{id}")
+//        public void deleteBooksForm(@PathVariable long id) {
+//            bookFormService.deleteById(id);
+//        }
+
         @DeleteMapping("/bookForm/{id}")
-        public void deleteBooksForm(@PathVariable long id) {
-            bookFormService.deleteById(id);
+        public String deleteBooksForm(@PathVariable Long id) {
+            if(bookFormService.existsById(id)) {
+                bookFormService.deleteById(id);
+                return "deleted";
+            }
+            return "bookForm do not exist in database!";
         }
 
         @PostMapping("/bookForm")
         public ResponseEntity<Object> createBookForm(@RequestBody BookForm bookForm) {
             BookForm savedBookForm = bookFormService.save(bookForm);
-
             URI location = ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}")
                     .buildAndExpand(savedBookForm.getId()).toUri();
-
             return ResponseEntity.created(location).build();
+        }
 
+        @PutMapping("/bookForm")
+        public BookForm updateBookForm(@RequestBody BookForm bookForm) {
+            return bookFormService.save(bookForm);
         }
